+++ save.go
@@ -23,6 +23,7 @@ import (
 	"fmt"
 	"io"
 	"runtime"
+	"sort"
 
 	"github.com/ctessum/geom"
 )
@@ -92,6 +93,17 @@ func (d *InMAP) initFromCells(cells []*Cell, emis *Emissions, config *VarGridCon
 	d.popIndices = make(map[string]int)
 	for i, p := range config.CensusPopColumns {
 		d.popIndices[p] = i
+	}
+	d.mortIndices = make(map[string]int)
+	mortRateColumns := make([]string, len(config.MortalityRateColumns))
+	i := 0
+	for _, m := range config.MortalityRateColumns {
+		mortRateColumns[i] = m
+		i++
+	}
+	sort.Strings(mortRateColumns)
+	for i, m := range mortRateColumns {
+		d.mortIndices[m] = i
 	}
 	for _, c := range cells {
 		d.InsertCell(c)